move integration tests into integration folder
