--- packages/perspective-viewer-d3fc/test/js/candlestick.spec.js
@@ -1,35 +1,0 @@
-/******************************************************************************
- *
- * Copyright (c) 2017, the Perspective Authors.
- *
- * This file is part of the Perspective library, distributed under the terms of
- * the Apache License 2.0.  The full license can be found in the LICENSE file.
- *
- */
-
-const path = require("path");
-
-const utils = require("@jpmorganchase/perspective-viewer/test/js/utils.js");
-
-const {withTemplate} = require("./simple-template");
-withTemplate("candlestick", "d3_candlestick", {template: "shares-template"});
-
-utils.with_server({}, () => {
-    describe.page(
-        "candlestick.html",
-        () => {
-            test.capture("filter by a single instrument.", async page => {
-                const viewer = await page.$("perspective-viewer");
-                await page.evaluate(element => element.setAttribute("filters", '[["Name", "==", "BARC"]]'), viewer);
-            });
-
-            test.capture("filter to date range.", async page => {
-                const viewer = await page.$("perspective-viewer");
-                await page.shadow_click("perspective-viewer", "#config_button");
-                await page.evaluate(element => element.setAttribute("column-pivots", '["Name"]'), viewer);
-                await page.evaluate(element => element.setAttribute("filters", '[["Date", ">", "2019-01-01"]]'), viewer);
-            });
-        },
-        {reload_page: false, root: path.join(__dirname, "..", "..")}
-    );
-});
